+++ src/elements/global/Header.js
@@ -1,10 +1,11 @@
+import Link from "../components/Link";
 
 function Header({urlImage, children}) {
     return (
         <header className="header">
-            <a aria-label="Little Lemon" href="./">
+            <Link href="./" variant={'link'} aria-label="Little Lemon" >
                 <img className='logo' src={urlImage} alt=""/>
-            </a>
+            </Link>
             {children}
         </header>
     );
